Add files via upload
Grootan Technologies-Coding Task   1. Create a mock API, which will return a list of users. 2. Have a dummy login and registration page. (You can use localStorage if needed) 3. Consume the API (created in Step 1) and get the list. 4. Display the records in the UI. 5. When a user clicks on a particular record, he should be able to see more details of that user. 6. While showing the details provide an option to view details of the next and previous user in the list. 7. Share the final source code via github 8. Deploy the frontend and backend(if any) to any free hosting service